(REF) SeatNowComponent 파일 분리 완료 (#20)
