+++ app/src/main/java/com/gmg/seatnow/presentation/owner/signup/steps/Step5PhotoScreen.kt
@@ -8,7 +8,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gmg.seatnow.presentation.component.AddPhotoButton
