+++ app/src/main/java/com/gmg/seatnow/presentation/owner/store/mypage/account/AccountInfoScreen.kt
@@ -7,7 +7,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.gmg.seatnow.presentation.component.SeatNowTopAppBar
