+++ app/src/main/java/com/gmg/seatnow/presentation/owner/store/withdraw/OwnerWithdrawScreen.kt
@@ -1,13 +1,11 @@
 package com.gmg.seatnow.presentation.owner.store.withdraw
 
-import android.R.id.bold
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Check
