+++ app/src/main/java/com/gmg/seatnow/presentation/user/mypage/UserMyPageScreen.kt
@@ -12,7 +12,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.gmg.seatnow.presentation.component.SeatNowMenuItem
 import com.gmg.seatnow.presentation.theme.SubGray
