+++ app/src/main/java/com/gmg/seatnow/presentation/user/mypage/UserWithdrawScreen.kt
@@ -1,7 +1,6 @@
 package com.gmg.seatnow.presentation.user.mypage
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
